Extract SelectCoinsForRound to CoinJoinCoinSelector static class
